update key pools
