+++ package.json
@@ -16,7 +16,7 @@
     "@oraichain/ethereum-multicall": "^1.0.2",
     "@oraichain/kawaiiverse-txs": "^0.0.3",
     "@oraichain/oraidex-common-ui": "^1.0.3",
-    "@oraichain/oraidex-contracts-sdk": "^1.0.24",
+    "@oraichain/oraidex-contracts-sdk": "^1.0.26",
     "@oraichain/oraidex-universal-swap": "^1.0.32",
     "@reduxjs/toolkit": "^1.9.3",
     "@sentry/react": "^7.47.0",
@@ -54,7 +54,7 @@
     "@oraichain/babel-plugin-operator-overloading": "^1.0.2",
     "@oraichain/common-contracts-build": "1.0.34",
     "@oraichain/cw-simulate": "^2.8.71",
-    "@oraichain/oraidex-contracts-build": "^1.0.20",
+    "@oraichain/oraidex-contracts-build": "^1.0.22",
     "@sentry/webpack-plugin": "^1.20.0",
     "@types/classnames": "^2.2.10",
     "@types/jest": "^26.0.14",
