+++ src/pages/Pools/components/Earning/Earning.tsx
@@ -12,12 +12,12 @@ import CosmJs from 'libs/cosmjs';
 import { getUsd } from 'libs/utils';
 import { isEqual } from 'lodash';
 import { useGetPoolDetail, useGetRewardInfo } from 'pages/Pools/hookV3';
-import { useGetStakingAssetInfo } from 'pages/Pools/hooks/useGetStakingAssetInfo';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { Type, WithdrawMining, fetchTokenInfo, generateMiningMsgs } from 'rest/api';
+import { Type, fetchTokenInfo, generateMiningMsgs } from 'rest/api';
 import styles from './Earning.module.scss';
 import { TokenItemType, ORAI } from '@oraichain/oraidex-common';
+import { WithdrawLP } from 'types/pool';
 
 type TokenItemTypeExtended = TokenItemType & {
   amount: bigint;
@@ -29,25 +29,15 @@ export const Earning = ({ onLiquidityChange }: { onLiquidityChange: () => void }
   const [address] = useConfigReducer('address');
   const [cachePrices] = useConfigReducer('coingecko');
   const [pendingRewards, setPendingRewards] = useState<TokenItemTypeExtended[]>([]);
-  const [stakingToken, setStakingToken] = useState<TokenItemType>();
   const [actionLoading, setActionLoading] = useState(false);
   const poolDetailData = useGetPoolDetail({ pairDenoms: poolUrl });
   const { info } = poolDetailData;
   const xOCH_PRICE = 0.4;
 
-  useEffect(() => {
-    if (!poolDetailData) return;
-    if (poolDetailData?.token1?.name === 'ORAI') {
-      setStakingToken(poolDetailData.token2);
-    } else {
-      setStakingToken(poolDetailData.token1);
-    }
-  }, [poolDetailData]);
-
-  const stakingAssetInfo = useGetStakingAssetInfo();
+  const liquidityToken = poolDetailData?.info?.liquidityAddr
   const { totalRewardInfoData, refetchRewardInfo } = useGetRewardInfo({
     stakerAddr: address,
-    assetInfo: stakingAssetInfo
+    stakingToken: liquidityToken
   });
 
   useEffect(() => {
@@ -117,8 +107,8 @@ export const Earning = ({ onLiquidityChange }: { onLiquidityChange: () => void }
       const msg = generateMiningMsgs({
         type: Type.WITHDRAW_LIQUIDITY_MINING,
         sender: address,
-        assetToken: stakingToken
-      } as WithdrawMining);
+        lpAddress: liquidityToken
+      } as WithdrawLP);
 
       const result = await CosmJs.execute({
         address: msg.contractAddress,
