+++ src/pages/Pools/components/Header/Header.tsx
@@ -25,11 +25,11 @@ export const useGetOraiPrice = () => {
       (pool) =>
         pool.firstAssetInfo === JSON.stringify(ORAI_INFO) &&
         pool.secondAssetInfo ===
-          JSON.stringify({
-            token: {
-              contract_addr: USDT_CONTRACT
-            }
-          })
+        JSON.stringify({
+          token: {
+            contract_addr: USDT_CONTRACT
+          }
+        })
     );
     if (!oraiUsdtPool) return;
     const oraiPrice = toDecimal(BigInt(oraiUsdtPool.askPoolAmount), BigInt(oraiUsdtPool.offerPoolAmount));
@@ -58,11 +58,11 @@ export const Header: FC = () => {
         .filter((rewardInfo) => rewardInfo.pending_reward !== '0')
         .map(
           (rewardInfo) =>
-            ({
-              contractAddress: network.staking,
-              msg: { withdraw: { asset_info: rewardInfo.asset_info } },
-              funds: null
-            } as ExecuteInstruction)
+          ({
+            contractAddress: network.staking,
+            msg: { withdraw: { staking_token: rewardInfo.staking_token } },
+            funds: null
+          } as ExecuteInstruction)
         );
 
       const result = await CosmJs.executeMultiple({
