+++ src/pages/Pools/components/StakeLPModal/StakeLPModal.tsx
@@ -16,13 +16,12 @@ import { FC, useEffect, useState } from 'react';
 import NumberFormat from 'react-number-format';
 import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import { generateMiningMsgsV3, Type } from 'rest/api';
+import { generateMiningMsgs, Type } from 'rest/api';
 import { RootState } from 'store/configure';
 import { ModalProps } from '../MyPoolInfo/type';
 import styles from './StakeLPModal.module.scss';
 import { ORAI, toDisplay, toAmount } from '@oraichain/oraidex-common';
 import { Pairs } from 'config/pools';
-import { useGetStakingAssetInfo } from 'pages/Pools/hooks/useGetStakingAssetInfo';
 
 const cx = cn.bind(styles);
 
@@ -33,7 +32,6 @@ export const StakeLPModal: FC<ModalProps> = ({
   myLpBalance,
   myLpUsdt,
   onLiquidityChange,
-  assetToken
 }) => {
   let { poolUrl } = useParams();
   const lpPools = useSelector((state: RootState) => state.token.lpPools);
@@ -43,10 +41,10 @@ export const StakeLPModal: FC<ModalProps> = ({
   const { info: pairInfoData } = poolDetail;
   const { lpTokenInfoData } = useGetPairInfo(poolDetail);
 
-  const stakingAssetInfo = useGetStakingAssetInfo();
+  const liquidityToken = poolDetail?.info?.liquidityAddr
   const { refetchRewardInfo } = useGetRewardInfo({
     stakerAddr: address,
-    assetInfo: stakingAssetInfo
+    stakingToken: liquidityToken
   });
   const [bondAmount, setBondAmount] = useState<bigint | null>(null);
   const [bondAmountInUsdt, setBondAmountInUsdt] = useState(0);
@@ -70,17 +68,14 @@ export const StakeLPModal: FC<ModalProps> = ({
     setActionLoading(true);
     displayToast(TToastType.TX_BROADCASTING);
     try {
+
       const oraiAddress = await handleCheckAddress('Oraichain');
       // generate bonding msg
-      const msg = generateMiningMsgsV3({
+      const msg = generateMiningMsgs({
         type: Type.BOND_LIQUIDITY,
         sender: oraiAddress,
         amount: parsedAmount.toString(),
-        lpAddress: lpTokenInfoData.contractAddress!,
-        assetInfo: Pairs.getStakingAssetInfo([
-          JSON.parse(pairInfoData.firstAssetInfo),
-          JSON.parse(pairInfoData.secondAssetInfo)
-        ])
+        lpAddress: liquidityToken
       });
 
       // execute msg
