+++ src/pages/Pools/helpers.ts
@@ -1,14 +1,6 @@
 import { fromBinary, toBinary } from '@cosmjs/cosmwasm-stargate';
-import { MulticallReadOnlyInterface } from '@oraichain/common-contracts-sdk';
-import {
-  Cw20Coin,
-  Asset,
-  AssetInfo,
-  InstantiateMarketingInfo,
-  OraiswapPairTypes,
-  OraiswapStakingTypes,
-  PairInfo
-} from '@oraichain/oraidex-contracts-sdk';
+import { Cw20Coin, MulticallReadOnlyInterface } from '@oraichain/common-contracts-sdk';
+import { Asset, AssetInfo, OraiswapPairTypes, OraiswapStakingTypes, PairInfo } from '@oraichain/oraidex-contracts-sdk';
 import { assetInfoMap, tokenMap, oraichainTokens } from 'config/bridgeTokens';
 import { ORAI, ORAIXOCH_INFO, SEC_PER_YEAR, STABLE_DENOM } from '@oraichain/oraidex-common';
 import { network } from 'config/networks';
@@ -34,6 +26,7 @@ import {
 } from '@oraichain/oraidex-common';
 import { Pairs } from 'config/pools';
 import { MinterResponse } from '@oraichain/oraidex-contracts-sdk/build/OraiswapToken.types';
+import { InstantiateMarketingInfo } from '@oraichain/common-contracts-sdk/build/Cw20Base.types';
 
 export type PairInfoData = {
   pair: PairInfoExtend;
@@ -90,12 +83,11 @@ export const calculateAprResult = (
 // Fetch APR
 const fetchAprResult = async (pairs: PairInfo[], pairInfos: PairInfoData[], prices: CoinGeckoPrices<string>) => {
   const lpTokens = pairs.map((p) => ({ contractAddress: p.liquidity_token } as TokenItemType));
-  const assetTokens = Pairs.getStakingInfoTokenItemTypeFromPairs(pairs);
   try {
     const [allTokenInfo, allLpTokenAsset, allRewardPerSec] = await Promise.all([
       fetchTokenInfos(lpTokens),
-      fetchAllTokenAssetPools(assetTokens),
-      fetchAllRewardPerSecInfos(assetTokens)
+      fetchAllTokenAssetPools(pairs),
+      fetchAllRewardPerSecInfos(pairs)
     ]);
     return calculateAprResult(pairs, pairInfos, prices, allTokenInfo, allLpTokenAsset, allRewardPerSec);
   } catch (error) {
@@ -220,17 +212,11 @@ export const calculateBondLpPools = (pairs: PairInfo[], res: AggregateResult) =>
 
 const generateRewardInfoQueries = (pairs: PairInfoExtend[], stakerAddress: string) => {
   const queries = pairs.map((pair) => {
-    let assetToken = assetInfoMap[pair.asset_infos_raw[0]];
-    const firstParsedAssetInfo = parseAssetInfo(pair.asset_infos[0]);
-
-    // we implicitly set asset info of the pool as non-ORAI token. If the first asset info in the pair list is ORAI then we get the other asset info
-    if (firstParsedAssetInfo === ORAI) assetToken = assetInfoMap[pair.asset_infos_raw[1]];
-    const { info: assetInfo } = parseTokenInfo(assetToken);
     return {
       address: network.staking,
       data: toBinary({
         reward_info: {
-          asset_info: assetInfo,
+          staking_token: pair.liquidity_token,
           staker_addr: stakerAddress
         }
       })
