+++ src/pages/Pools/index.module.scss
@@ -2,320 +2,83 @@
 @import 'src/styles/themes';
 
 .pools {
-  padding: 34px;
-  color: #ffffff;
-  height: 100%;
+    padding: 34px;
+    margin: 0 auto;
+    height: 100%;
+    max-width: 90%;
 
-  @media (max-width: 575px) {
-    padding: 0;
-  }
-}
-
-.header {
-  &_title {
-    font-weight: 600;
-    font-size: 40px;
-
-    @media (max-width: 575px) {
-      font-size: 30px;
+    @include ultra_desktop_max() {
+        max-width: 100%;
+        padding: 0;
+        margin: 24px 0;
     }
 
-    @include theme() {
-      color: theme-get('text-color');
-    }
-  }
-
-  &_data {
-    margin-top: 15px;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    gap: 10px;
-
-    &_item {
-      margin-right: 80px;
-
-      @media (max-width: 575px) {
-        margin: 0;
-        width: 100%;
-      }
+    @include mobile() {
+        max-width: 100%;
+        padding: 0;
+        margin: 24px 0;
     }
 
-    &_name {
-      font-weight: 600;
-      font-size: 24px;
-      color: #777e90;
-
-      @media (max-width: 575px) {
-        font-size: 16px;
-      }
+    @include medium-desktop() {
+        max-width: 100%;
     }
 
-    &_value {
-      font-weight: 600;
-      font-size: 24px;
-      margin-left: 10px;
-
-      @media (max-width: 575px) {
-        font-size: 20px;
-      }
-
-      @include theme() {
-        color: theme-get('text-color');
-      }
+    @include small-mobile {
+        margin-top: 85px;
     }
-  }
-}
-
-.watchlist {
-  margin-top: 25px;
-
-  &_title {
-    font-weight: 600;
-    font-size: 18px;
-    color: #777e90;
-  }
-}
-
-.pairbox {
-  @include theme() {
-    background: theme-get('bg-pair');
-    box-shadow: theme-get('shadow-pair');
-  }
-
-  border-radius: 24px;
-  padding: 24px 24px 35px;
-  width: calc(25% - 27px);
-  border: 1px solid;
-  border-image-source: linear-gradient(
-    99.99deg,
-    rgba(255, 255, 255, 0) 0%,
-    rgba(255, 255, 255, 0.15) 26.06%,
-    rgba(225, 240, 255, 0.202656) 64.17%,
-    rgba(232, 241, 255, 0) 98.89%
-  );
-
-  @media (max-width: 1600px) {
-    width: calc((100% / 3) - 24px);
-  }
-
-  @media (max-width: 1200px) {
-    width: calc(50% - 18px);
-  }
-
-  @media (max-width: 600px) {
-    width: calc(100%);
-  }
-
-  &:hover {
-    cursor: pointer;
-    border: 1px solid #a871df;
-  }
-
-  &_header {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    align-items: center;
-  }
-
-  &_logos {
-    position: relative;
-    width: 100px;
-    height: 60px;
-  }
-
-  &_logo1 {
-    position: absolute;
-    left: 0;
-    top: 0;
-    z-index: 0;
-    width: 48px;
-    height: 48px;
-  }
-
-  &_logo2 {
-    position: absolute;
-    left: 25px;
-    top: 30px;
-    z-index: 1;
-    width: 48px;
-    height: 48px;
-  }
-
-  &_pair_name {
-    font-weight: 600;
-    font-size: 22px;
-
-    @include theme() {
-      color: theme-get('text-color');
-    }
-  }
-
-  &_pair_rate {
-    margin: 5px 0;
-    color: #777e90;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 150%;
-  }
-
-  &_pair_apr {
-    background: rgba(97, 47, 202, 0.5);
-    border-radius: 4px;
-    font-weight: 600;
-    font-size: 14px;
-    padding: 4px 6px;
-  }
-
-  &_content {
-    margin-top: 40px;
-  }
-
-  &_data {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-
-    &_name {
-      font-weight: 500;
-      font-size: 14px;
-      line-height: 150%;
-      color: #777e90;
-    }
-
-    &_value {
-      font-weight: 600;
-      font-size: 16px;
-
-      @include theme() {
-        color: theme-get('text-color');
-      }
-    }
-  }
 }
 
 .listpoolsToken {
-  &_create {
-    display: flex;
-    align-items: center;
-    @media (max-width: 930px) {
-      padding-top: 10px;
+    &_create {
+        display: flex;
+        align-items: center;
+
+        @include mobile {
+            padding-top: 10px;
+        }
+
+        .create_item {
+            height: 48px;
+            padding: 8px 16px;
+            font-size: 16px;
+            font-weight: 500;
+            cursor: pointer;
+            margin-right: 15px;
+            background: #1e1e21;
+            border-radius: 8px;
+            display: inline-flex;
+            align-items: center;
+            justify-content: center;
+
+            &:hover {
+                background: #804af2 !important;
+            }
+        }
     }
-    .create_item {
-      height: 48px;
-      padding: 8px 16px;
-      font-size: 16px;
-      font-weight: 500;
-      cursor: pointer;
-      margin-right: 15px;
-      background: #1e1e21;
-      border-radius: 8px;
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      &:hover {
-        background: #804af2 !important;
-      }
-    }
-  }
 }
 
-.listpools {
-  margin-top: 40px;
-  .listpools_all {
-    display: flex;
-    justify-content: space-between;
-    @media (max-width: 930px) {
-      flex-direction: column;
-      justify-content: space-between;
-    }
-  }
-  .no_data {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    border-radius: 16px;
-    padding: 60px 20px;
-
-    @include theme() {
-      color: theme-get('bg-color');
-    }
-
-    span {
-      color: #505665;
-      font-size: 20px;
-      font-weight: 400;
-      margin-top: -20px;
-    }
-  }
-
-  &_header {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-  }
-
-  &_filter {
+.symbols {
     display: flex;
     align-items: center;
 
-    .filter_item {
-      height: 48px;
-      padding: 8px 16px;
-      font-size: 16px;
-      font-weight: 500;
-      cursor: pointer;
-      margin-right: 15px;
+    &_logo {
+        margin-right: 12px;
 
-      @include theme() {
-        background-color: theme-get('bg-color');
-        color: theme-get('text-color-filter');
-      }
-
-      &:hover {
-        @include theme() {
-          background-color: theme-get('bg-hover-color');
-          color: $white-color;
+        &_right {
+            position: relative;
+            right: 10px;
+            width: 44px;
+            height: 44px;
         }
-      }
 
-      border-radius: 8px;
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
+        &_left {
+            width: 44px;
+            height: 44px;
+        }
     }
 
-    .filter_active {
-      color: $blue-color !important;
-
-      &:hover {
-        color: $white-color !important;
-      }
+    &_name {
+        font-weight: 500;
+        font-size: 16px;
     }
-  }
-
-  &_search {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    @media (max-width: 575px) {
-      margin-top: 10px;
-      width: 100%;
-    }
-  }
-
-  &_create {
-    padding: 16px 76px;
-  }
-
-  &_list {
-    display: flex;
-    gap: 35px;
-    margin-top: 35px;
-    flex-wrap: wrap;
-  }
 }
